i now know understand the basic stuff on the latex code and have done some comments on what i think each line does
for this i had to reseach what some keywords do (e.g. super and __init__).
me and danish have also seen the cpp and base_gem and also have a basic knowledge of that but we didnt comment on it like i have do eon hte latex one.

danish is not in today, but you probably already know that.

i will also try to improve my understanding.
i have also saw the unity program which is fairly similar but will add anything thats different and complete this by 3 30